backend/requirements.txt
fastapi~=0.110
uvicorn[standard]~=0.29
sentence-transformers~=2.7
faiss-cpu~=1.8
numpy~=1.26
orjson~=3.10
pydantic~=2.7
langdetect~=1.0.9
tqdm~=4.66
python-dotenv~=1.0
pytest~=8.0
slowapi~=0.1.9
prometheus-fastapi-instrumentator~=6.1
pdfplumber>=0.10.3